+++ src/components/imgui_widgets/ImGuiWidgetDeviceInfo.cpp
@@ -1,0 +1,6 @@
+#include "ImGuiWidget.h"
+
+void ImGuiWidgetDeviceInfo::Draw(const VulkanEngine* engine) {
+
+
+}
